Streamline test imports by replacing absolute paths with relative paths.
