+++ src/qtm4j/tool/test-execution/link-test-case-bugs.test.ts
@@ -1,11 +1,11 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import { ENDPOINTS } from "../../../../../qtm4j/config/constants";
-import { ResolverKeys } from "../../../../../qtm4j/config/field-resolution.types";
+import { ENDPOINTS } from "../../config/constants";
+import { ResolverKeys } from "../../config/field-resolution.types";
 import {
   LinkBugsResponse,
   LinkTestCaseBugsBody,
-} from "../../../../../qtm4j/schema/link-bugs.schema";
-import { LinkTestCaseBugs } from "../../../../../qtm4j/tool/test-execution/link-test-case-bugs";
+} from "../../schema/link-bugs.schema";
+import { LinkTestCaseBugs } from "./link-test-case-bugs";
 
 describe("LinkTestCaseBugs", () => {
   let mockClient: any;
